PP-5991: Split working code into Component files; package build
* individual component files are simpler to read and significantly
speed up the compiler/ linting in IDE
* Typescript is watching all files outside of the tree, split folders into
browser and server (src) until webpack is used for both.
* integrate browser build into main package scripts
