+++ package.json
@@ -15,15 +15,15 @@
     "test": "npm run test:unit",
     "test:unit": "AUTH_GITHUB_ENABLED=true mocha -r ts-node/register -r dotenv/config --recursive \"./src/**/*.spec.{js,ts}\" dotenv_config_path=src/tests/.env",
     "test:cypress": "./scripts/run-cypress",
-    "build": "scripts/build && npm run build:typescript && npm run build:copy && npm run build:sass",
     "compile": "npm run build",
+    "build": "scripts/build && npm run build:typescript && npm run build-browser && npm run build:copy && npm run build:sass",
     "build-browser": "webpack --mode production",
     "build-browser-dev": "webpack -w --mode development",
     "build:typescript": "tsc",
     "build:sass": "scripts/compile-assets",
     "build:copy": "copyfiles --up 1 src/**/*.njk dist/",
     "preinstall": "echo \"Preinstall hook\" && exit 0",
-    "prepare": "npm run build:sass",
+    "prepare": "npm run build:sass && npm run build-browser",
     "postinstall": "echo \"postinstall set to run with prepare (only in dev)\" && exit 0",
     "version": "echo $npm_package_version"
   },
